Minor fixes

+++ DES/DES.cs
@@ -149,7 +149,7 @@ namespace DES
             return block1.Zip(block2, (block1Bit, block2Bit) => block1Bit ^ block2Bit).ToArray();
         }
 
-        private static IEnumerable<bool> STransform(bool[] block, IEnumerable<int> sTable)
+        private static IEnumerable<bool> STransform(bool[] block, IEnumerable<byte> sTable)
         {
             var row = Convert.ToInt16(string.Concat(FromBool(block[0]), FromBool(block[5])), 2);
             var col = Convert.ToInt16(string.Concat(block.Skip(1).Take(4).Select(FromBool)), 2);
